feat: create item carousel
